requirements.txt
absl-py==2.2.2
altair==5.5.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.1
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.21.0
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.33.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.3
numpy==2.1.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.2.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.21
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.3
pytest==7.4.4
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.17.1
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
terminado==0.18.0
thop==0.1.1.post2209072238
tinycss2==1.2.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.101
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.2.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wrapt==1.17.2
